HomeScreen RunScreen navigation + Refactor
* Upgraded react-native from 0.69.5 to 0.69.6
* Time format refactoring
* HomeScreen RunScreen navigation

Dropdown.js
* style refactoring

RunScreen.js
* Refactoring
* Added alerts to action button

+++ screens/HomeScreen.js
@@ -1,31 +1,31 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { View, Text, TouchableOpacity, StyleSheet, FlatList } from "react-native";
 import TaskComp from "../components/TaskComp";
 import { createNewTask } from "../Utils";
 
-const _tranddur = () => Math.floor(Math.random()*300)+180; // temp random duration
-
 const tasks_list = [
-    createNewTask('Pushup'),
     createNewTask('Run'),
-    createNewTask('Squat'),
+    createNewTask('Squats'),
     createNewTask('Walk'),
-    createNewTask('Streching'),
     createNewTask('Plank'),
-    createNewTask('Walk'),
 ]
 
-function useForceUpdate(){
+function useForceUpdate() {
     const [value, setValue] = useState(0);
     return () => setValue(value => value + 1); // update state to force render
-
 }
 
-const HomeScreen = ({navigation}) => {
+const HomeScreen = ({ navigation, route: { params } }) => {
     const forceUpdate = useForceUpdate();
 
-    
-    
+    if (params && params.taskItem) {
+        tasks_list.push(params.taskItem);
+        navigation.setParams({ taskItem: undefined });
+        return;
+    }
+
+
+
     const tasks_editor = {
         deleteItem: (n) => {
             tasks_list.splice(n, 1);
@@ -36,17 +36,16 @@ const HomeScreen = ({navigation}) => {
     return (
         <View style={styles.container}>
             <FlatList
-                style={{flex:1}}
+                style={{ flex: 1 }}
                 contentContainerStyle={{ paddingVertical: 10 }}
                 data={tasks_list}
-                renderItem={(t) => <TaskComp task={t} editor={tasks_editor}/>}
+                renderItem={(t) => <TaskComp task={t} editor={tasks_editor} />}
             />
-            <TouchableOpacity activeOpacity={1} style={styles.add_btn}
-            onPress={() => {
-                //tasks_list.push(createNewTask('Pushup'));
-                //forceUpdate();
-                navigation.navigate('Run');
-            }}>
+            <TouchableOpacity
+                activeOpacity={1}
+                style={styles.add_btn}
+                onPress={() => navigation.navigate('Run')}
+            >
                 <Text style={styles.add_btn_text}>+</Text>
             </TouchableOpacity>
         </View>
@@ -58,7 +57,7 @@ const styles = StyleSheet.create({
         flex: 1
     },
     add_btn: {
-        height:50,
+        height: 50,
         alignItems: 'center',
         justifyContent: 'center',
         backgroundColor: '#fff',
